# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface CatalogosResponse { [key: string]: unknown }

export interface ResumenResponse { [key: string]: unknown }

export interface HealthStatus {
  status: string;
}

export interface LoginRequest {
  email: string;
  /** @minLength 1 */
  password: string;
}

export type AppRole = typeof AppRole[keyof typeof AppRole];


export const AppRole = {
  gerencia: 'gerencia',
  gerente_comercial: 'gerente_comercial',
  coordinador: 'coordinador',
  asesor: 'asesor',
} as const;

export interface AuthUser {
  id: string;
  email: string;
}

export interface AuthProfile {
  id: string;
  email: string;
  /** @nullable */
  nombreCompleto: string | null;
  /** @nullable */
  sucursalId: string | null;
  /** @nullable */
  unidadNegocioId: string | null;
  isAdmin: boolean;
  unidadesNegocioIds: string[];
  sucursalesIds: string[];
}

export interface AuthPayload {
  user: AuthUser;
  profile: AuthProfile;
  role: AppRole | null;
}

export interface Problem {
  message: string;
}

export type Logout200 = {
  success: boolean;
};

export type GetResumenParams = {
anio: number;
/**
 * Comma-separated month numbers, or all
 */
meses?: string;
sucursalId?: string;
unidadNegocioId?: string;
};