# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Session authentication
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Start a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authenticated user and scope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthPayload"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
        "429":
          description: Too many attempts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get the current session
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Current authenticated user and scope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthPayload"
        "401":
          description: Missing or expired session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: End the current session
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Session ended
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required: [success]
  /catalogos:
    get:
      operationId: getCatalogos
      tags: [data]
      summary: Get commercial catalogs
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Branches and business units available to the current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogosResponse"
        "401":
          description: Missing or expired session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
  /resumen:
    get:
      operationId: getResumen
      tags: [data]
      summary: Get the commercial summary
      security:
        - cookieAuth: []
      parameters:
        - name: anio
          in: query
          required: true
          schema:
            type: integer
        - name: meses
          in: query
          required: false
          description: Comma-separated month numbers, or all
          schema:
            type: string
        - name: sucursalId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: unidadNegocioId
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Commercial summary data scoped to the current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumenResponse"
        "401":
          description: Missing or expired session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: ccv_session
  schemas:
    CatalogosResponse:
      type: object
      additionalProperties: true
    ResumenResponse:
      type: object
      additionalProperties: true
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
      required: [email, password]
    AppRole:
      type: string
      enum: [gerencia, gerente_comercial, coordinador, asesor]
    AuthUser:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
      required: [id, email]
    AuthProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        nombreCompleto:
          type: [string, "null"]
        sucursalId:
          type: [string, "null"]
          format: uuid
        unidadNegocioId:
          type: [string, "null"]
          format: uuid
        isAdmin:
          type: boolean
        unidadesNegocioIds:
          type: array
          items:
            type: string
            format: uuid
        sucursalesIds:
          type: array
          items:
            type: string
            format: uuid
      required: [id, email, nombreCompleto, sucursalId, unidadNegocioId, isAdmin, unidadesNegocioIds, sucursalesIds]
    AuthPayload:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/AuthUser"
        profile:
          $ref: "#/components/schemas/AuthProfile"
        role:
          anyOf:
            - $ref: "#/components/schemas/AppRole"
            - type: "null"
      required: [user, profile, role]
    Problem:
      type: object
      properties:
        message:
          type: string
      required: [message]